# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/google-contacts/oauth/authorize-url": {
      "post": {
        "operationId": "GoogleContactsController_createAuthorizeUrl",
        "summary": "Create the Google OAuth authorization URL",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Authorization URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorizeUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "google-contacts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/agent-internal/google-contacts/sync-contact": {
      "post": {
        "operationId": "GoogleContactsInternalController_syncContact",
        "summary": "Sync a WhatsApp contact to Google Contacts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncGoogleContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact sync result"
          }
        },
        "tags": [
          "google-contacts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/agent-internal/products/sample": {
      "get": {
        "operationId": "ProductsInternalController_getSampleProducts",
        "summary": "Récupérer un échantillon de produits",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent pour récupérer un échantillon représentatif du catalogue (génération/mise à jour du prompt de description image). Non destiné au frontend.",
        "parameters": [
          {
            "name": "max",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perCollection",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Échantillon de produits de l'agent retourné"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent-internal/products/by-retailer-id/{retailerId}": {
      "get": {
        "operationId": "ProductsInternalController_getProductByRetailerId",
        "summary": "Trouver un produit par retailer_id",
        "description": "Endpoint interne backend, appelé par le pipeline image du whatsapp-agent après OCR pour matcher rapidement un produit via son code retailer.",
        "parameters": [
          {
            "name": "retailerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit correspondant retourné (ou null)"
          },
          "400": {
            "description": "retailerId manquant ou invalide"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent-internal/products/by-id/{productId}": {
      "get": {
        "operationId": "ProductsInternalController_getProductByAnyId",
        "summary": "Trouver un produit par identifiant interne/WhatsApp/retailer",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent pour résoudre un identifiant produit vers le produit métier et son whatsapp_product_id.",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit correspondant retourné (ou null)"
          },
          "400": {
            "description": "productId manquant ou invalide"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent-internal/products/by-ids": {
      "get": {
        "operationId": "ProductsInternalController_getProductsByAnyIds",
        "summary": "Trouver plusieurs produits par identifiants interne/WhatsApp/retailer",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent pour résoudre plusieurs identifiants produit et récupérer les données nécessaires à la construction d'une preview de lien produit.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste ordonnée des correspondances retournée"
          },
          "400": {
            "description": "Aucun identifiant valide fourni"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent-internal/products/search-by-keywords": {
      "get": {
        "operationId": "ProductsInternalController_searchProductsByKeywords",
        "summary": "Rechercher des produits par mots-clés",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent pour la recherche OCR textuelle. Le matching est strictement effectué sur retailer_id.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Produits et mots-clés matchés retournés"
          },
          "400": {
            "description": "Aucun mot-clé valide fourni"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent-internal/products/cover-image-descriptions": {
      "patch": {
        "operationId": "ProductsInternalController_batchUpdateProductImageIndexing",
        "summary": "Mettre à jour en batch les descriptions de cover et états d'indexation",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent en fin d'indexation pour persister en une seule requête les descriptions de cover et remettre les flags d'indexation à false (produits et images).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdateProductImageIndexingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch appliqué sur les produits appartenant à cet agent"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/agent-internal/products/for-image-indexing": {
      "get": {
        "operationId": "ProductsInternalController_getProductsForImageIndexing",
        "summary": "Lister les produits pour l'indexation image",
        "description": "Endpoint interne backend, appelé par le whatsapp-agent pour récupérer les produits et leurs images de couverture à indexer dans Qdrant (traitement fait côté agent).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste de produits prêts pour l'indexation image"
          },
          "401": {
            "description": "JWT inter-services invalide ou absent"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SyncGoogleContactDto": {
        "type": "object",
        "properties": {}
      },
...
      "BatchUpdateProductImageIndexingDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}